Day 23 Problem 2 with huge props to https://raw.githack.com/ypsu/experiments/master/aoc2018day23/vis.html

+++ Day 23/Nanobot.java
@@ -1,8 +1,8 @@
 public class Nanobot
 {
-  public final long radius;
+  public final int radius;
 
-  public Nanobot(long radius)
+  public Nanobot(int radius)
   {
     this.radius = radius;
   }
